+++ lib/src/callbacks/nyx_session_callback.dart
@@ -1,0 +1,1 @@
+typedef NyxSessionCallback = void Function(int sessionId);